requirements.txt
amqp==5.0.5
anyjson==0.3.3
arrow==0.14.7
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-events-rest-framework==0.1
django-rest-framework==0.1.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
dnspython==1.16.0
eventlet==0.30.1
greenlet==1.0.0
humanize==3.2.0
ics==0.7
idna==2.10
kombu==5.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.23
sqlparse==0.4.1
TatSu==5.5.0
tornado==6.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
amqp==5.0.5
anyjson==0.3.3
arrow==0.14.7
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-events-rest-framework==0.1
django-rest-framework==0.1.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
dnspython==1.16.0
eventlet==0.30.1
greenlet==1.0.0
gunicorn==20.0.4
humanize==3.2.0
ics==0.7
idna==2.10
kombu==5.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.23
sqlparse==0.4.1
TatSu==5.5.0
tornado==6.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
